ref-manual, dev-manual: Point out "xterm" needed for eclipse use.
Fixes [YOCTO #5766]

In the ref-manual, I added "Eclipse" use in the title for the
"Graphics Support" packages because the "xterm" package is required
for graphics support across all distributions.

In the dev-manual, I specifically called out the need for "xterm"
as a host package at the top of the section describing workflow
with Eclipse.  I also updated the cross-reference links to the
host package and Linux distro requirements to point into the
ref-manual, which is more comprehensive, than into the QS.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>