dev-manual, ref-manual: Added error reporting tool information
Fixes [YOCTO #1562]

In the dev-manual, I created a new section called "Using the
Error Reporting Tool" that describes how to use this feature.
The text was based on information I received from Andreea Proca.

In the ref-manual, I added a new variable entry for ERR_REPORT_DIR
in the glossary and a new report-error class in the class chapter.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>