ref-manual: Removed the BitBake chapter.
This information now exists in the separate BitBake User Manual.
To remove this information, the ref-bitbake.xml file was taken
out of the ref-manual.xml build, an overview bullet describing
BitBake chapter was removed from the "Introduction" chapter,
and one cross-reference link to a sub-section of the old BitBake
chapter had to be re-routed into the similar section of the BitBake
Manaul.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>