documentation: Updates to get rid of PRINC variable.
I did not eliminate this variable but rather changed the description
to indicate that it has been deprecated and that the user should now
use a PR Service to do this stuff.  There were several areas in the
YP manual set that used the variable in code.  I just deleted those
lines from the code.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>