ref-manual: Added useradd-staticids class and edits to some variables.
Added the new useradd-staticids class to the reference section
for classes.

Updated the USERADD_UID_TABLES and USERADD_GID_TABLES variables
to tell how the system uses BBPATH in the default mode to
get UID and GID information from the files/passwd and files/group
files.

Added a note indicating that you can delete the TMPDIR directory
to fix things up if you configure useradd-staticids in a
configured system.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>