dev-manual: Added "Properly Versioning Pre-Release Recipes" section.
Fixes [YOCTO #6243]

This section addresses situations when working with recipe
file names of recipes that have not been released (e.g.
recipe-name_1.0.3.rc1).  When the recipe is released the system
has difficulty recognizing the new recipe name as a later version.
So the work-around is to use PV to set the recipe version in
the recipe.  The section describes this.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>