ref-manual: Added new chapter on task documentation
Changes to support a new chapter on the 51 tasks that the
OpenEmbedded build system defines.  The changes include a new
file called ref-tasks.xml, changes to the ref-manual.xml
building file to include the new chapter, and a new bullet
item to the list that describes what is in the reference
manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>