ref-manual: Edits to the "Making Images More Secure" section.
Fixes [YOCTO #5482]

I added some key references to the section on considerations
specific to the OpenEmbedded build system.  In particular, I
provided some cross-linking back to the extrausers.bbclass
section to reference an example of adding a user account.  I
also split out the topics of adding an extra user and setting
a password on the image in the bulleted list.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>