template: Removed the "d:" character namespace prefix.
The prefix is used for DocBook 5 Schema.  We are using the
4.2 Schema.  Having this prefix disabled the permalink titles
XSL template (section.title.xsl) produced permalinks with no
titles.  Because I removed it in the section.title.xls file,
I also removed it here for completeness.  Note that I don't
think the template is used at this point but I made the change
just in case for the future.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>