ref-manual: Added ptest default build info to migrating section.
Added a ptest section to the "Moving to the Yocto Project 1.6 Release" section.

Fixes [YOCTO #6400]

Added a new section named "Package Tests (ptest)" to the migration
section stating that ptest is built but not installed by default.
I also added some cross-reference links to a usage section and to
the ptest.bbclass section.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>