dev-manual, yocto-project-qs: New section on working with source files.
Fixes [YOCTO #5566]

For the dev-manual, I created a new section called "Working
with Source Files."  In the section, I cover how to set up mirrors
and also how to pre-fetch source using the bitbake -c fetchall
<target> command.

For the yocto-project-qs, I removed the mirror information in the
"Super User" section, which became redundant with the new section
now in the dev-manual.  I also, removed the fetchall variation of
the bitbake command.  Both areas reference into the new section of
the dev-manual now.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>