mega-manual.sed: Fixed search string problem for profile-manual.
Found a very subtle problem with the search string that processes
links to the Yocto Project Profiling and Tracing Manual where the
links go to the top-level (i.e. no ID tag in the link).

I had the name of the manual as "Yocto Project Profile and
Tracing Manual", which means there would never be a match.
Consequently, when the Makefile called the mega-manual.sed file
to process the links in mega-manual.html, any top-level link
to that manual was not processed and was being left as a hard
link to the versioned manual.  Processing a top-link should
convert it to a non-link (for now).

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>

Conflicts:

	documentation/tools/mega-manual.sed
        Had to clean up some conflicts to get the cherry-pick
        to work.  It seems the line for the profile manual was
        not even in this sed file.  Also, had to reset the
        1.4.4 strings to 1.5.3.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>