ref-manual, dev-manual: Updates to support PTEST_ENABLED.
In the ref-manual, I added a new variable description for the
PTEST_ENABLED variable.

In the dev-manual, I added a note to the "Adding ptest to Your
Build" section to mention that the OpenEmbedded system uses
PTEST_ENABLED to enable or disable building ptests and that
the user does not set this variable directly.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>