mega-manual.sed: Updates to support links to BB manual
Discovered that the sed file was processing the bitbake-user-manual
just like the others, which was wrong.  the BB manual is not
included in the mega-manual so the links there need to be left
alone.  The fix included specially creating substitution lines
to handle all the YP docs while excluding the BB manual.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>