dev-manual: Modified how to find checksums in the recipes
The section on wrting a new recipe suggested using a "build-fail"
method to get bitbake to return exact checksums for code.  It
was pointed out that this could be a middle-security risk and
that we should not do that but instead get the signatures from
the upstream website.  However, many times those sites don't
provide that information.  I re-worded the text to note the
ideal method (upstream checksums) and then resorted to the
"build-fail" method as a "way" to also get them when the upstream
location does not have them.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>