ref-manual: INITRD var: make it a list of filesystem images.
The initrd image used by the Linux kernel is list of file
system images concatenated together and presented as a single
initrd file at boot time.

So far the initrd is a single filesystem image. But in cases
like to support early microcode loading, the initrd image need
to have multiple filesystem images concatenated together.

This commit is extending the INITRD variable from a single
filesystem image to a list of filesystem images to satisfy
the need mentioned above.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>