dev-manual: Misc edits to the model chapter (Chapter 4).
I replaced user-supplied variables with <replaceable></replaceable>
tags to get proper formatting.  I inserted several references
to the new QEMU chapter for help using QEMU.  I updated a note
on the BSP available so that it is a bit more generic and not so
dependent on an exact list within meta-intel and the source
repositories.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>