ref-manual: Updated module_conf variable
This description was a bit confusing with the way the general
syntax was described so I re-wrote it a bit.  Also, added the
requirement that if you use module_conf for a given <modname>,
then you must also include the <modname> as part of the new
KERNEL_MODULE_PROBECONF variable.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>