poky.ent: post release fix of the POKYVERSION_COMPRESSED variable.
Missed this one and it is used to resolve the YOCTO_RELEASE_NOTES
URL in the dev-manual and the ref-manual.  The value was left at
"1100" when it should have been "1200".  I changed it post-release.

This means that the tarball is bad but the HTML versions published on
the server are correct for dizzy.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>