ref-manual: Added info strings for variables in glossary.
Fixes [YOCTO #5600]

Added a <info></info> tag pair with a short description of the
variable to each variable in the glossary.  The short description
is what will be used in the documentation.conf file.  Scripting
from engineering will pick up the variable description
automatically.  The information between the tags does not
appear in the HTML or PDF output.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>