ref-manual: Created form for context images for variables
Fixes [YOCTO #6158]

Updated the style scheet file to format glossary variables in
such a way to allow for a PNG file of standard size to be
displayed below each variable.  Right now, the PNG file is a
generic placeholder.  Each variable displays a PNG file.

Updated each variable to include that PNG file.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>