yocto-project-qs: Removed redundant code to display title image.
Creation of a PDF version of the YP Quick Start is not supported.
The <mediaobject></mediaobject> code appears in the manual for
that purpose only.  It is ignored during HTML generation.  Thus,
I have removed it from the file.  The image is actually displayed
through the <authorgroup> definition, which is located in the
CSS stylesheet.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>