ref-manual: Updates to some tuning variables in the glossary:
Fixes [YOCTO #5463]

Updated the notes in the TUNE_ASARGS, TUNE_CCARGS, and
TUNE_LDARGS to indicate that the BSPs select the tune and then
selected tun affects the tun variables themselves.

Also combined the TUNE_PKGARCH and the TUNE_PKGARCH_tune variable.
We really don't need the TUNE_PKGARCH_tune variable as it is simply
the TUNE_PKGARCH variable with the override.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>