dev-manual: Updates to the "Using .bbappend Files" section.
This section was a bit confusing on the use of "_prepend" as associated
with the FILESEXTRAPATHS statement.  Especially in the final note
where it indicated you don't need "_prepend" if your append file
is not adding any new files.  I updated the note to be more specific.

Another small change I made was to replace the Crown Bay BSP referenced
in the section with the Emenlow BSP, which has a longer life span than
CB.

Reported-by: Robert P. J. Day <rpjday@crashcourse.ca>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>