dev-manual: Updated section on how checksums are dealt with
In the "Fetching Code" section, some loose description of how the
build system reacts to the two checksums in a recipe existed.
Robert asked some questions about how the behavior would be if
just a single, correct checksum existed and if the two existed but
only one proved correct.  Turns out, if you supply a single one then
the build will not throw an error or warning.  If you supply two
but one is incorrect, then the build will throw an error or
warning.  I updated the text to be technically correct without
going into huge details.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>