dev-manual, ref-manual, yocto-project-qs: Updates to parallelism variables
The way the old manual set had worded the use of the
BB_NUMBER_THREADS, PARALLEL_MAKE, and BB_NUMBER_PARSE_THREADS
had become depricated.  It suggested a lot of user intervention
to get these variables set to have the build system utilize
optimal parallelism.  The change a while back to the build system
where the variables are set to the number of build host cores
now makes the variables automatically be set to their best
values.  I have gone through the manual set and changed wordings
appropriately to reflect this.

Reported-by: Robert P. J. Day <rpjday@crashcourse.ca>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>