adt-manual: Expanded note with a user configuration reference.
The note in the "Getting the ADT Installer Tarball" section has
a note that warns the user to be sure they have local.conf
properly configured before baking out the adt-installer.  I
provided a linked cross-reference into the section of the
ref-manual that takes a closer look at user configuration.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>