adt-manual, ref-manual: Detail on using an external toolchain.
Fixes [YOCTO #4907]

I created a new section in the toolchain chapter of the adt-manual
called "optionally Using an External Toolchain".  This provides
some fundamental information for the user that wants to use
an external toolchain.  Additionally, it references Mentor
Graphics Sourcery toolchain as an example.

In the ref-manual, I updated the TCMODE variable description to
better integrate the MGC Sourcery toolchain as the overriding
example.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>