ref-manual: Updates to AUTOREV and SRCREV variable descriptions
Fixes [YOCTO #7306]

Added an example for when the user tries to inherit the most recent
version of software through use of the SRCREV variable.  When using
SRCREV = "${AUTOREV}" PV does not always get ${SRCPV}, which it need
in order to fetch the latest software.  I added a clarifying example
to explain this situation.

This commit also has a small change to fix a broken link to
EXTERNAL_TOOLCHAIN.  We do not documentat that variable so I removed
the link.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>