ref-manual: Repositioned PNG files for context image.
I moved the PNG file for the context to be the first item displayed
for a given glossary term.  The PNG file must be negatively indented
though to look decent.

Fixes [YOCTO #5601]

These reformatting changes were needed because of brower incompatibilities
discovered with the existing solution for displaying these PNG files
with glossary terms.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>