bsp-guide: Removed 3.10 and 3.17 kernel references.
Version 1.8 drops support of the 3.10 and 3.17 kernels.  The
emenlow example was using the 3.10 kernel recipe as a showcase
and it also listed the recipe in the directory structure.  I
removed it.  Additionally, the yocto-bsp create command was
prompting for the user to accept the default 3.17 kernel.  I changed
that to 3.19.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>