poky-ent: update OpenSUSE essentials
Replaced texinfo -> makeinfo in the list of essential packages for
OpenSUSE.

texinfo was in the essentials to only provide makeinfo utility.
/usr/bin/makeinfo is provided by makeinfo package on OpenSUSE distro, so
it's enough to install makeinfo to be able to build poky.

This change reduces number of packages installed on the host system as
installing texinfo causes installation of around 1500 texlive-*
packages and requires around 1Gb of additional disk space on OpenSUSE.

Signed-off-by: Ed Bartosh <ed.bartosh@linux.intel.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>