dev-manual: Removed inc PR step from Quilt workflow
Fixes [YOCTO #7460]

There was a single occurrence of specifically calling out that PR
needed to be manually incremented in the workflow section of the
manual.  It was step 8 of the Quilt flow.  I deleted this step.
Previous edits to the set have included extensive instruction on using
a PR Service, which is basically what replaces the need for manually
updating PR as it was in the past.

Signed-off-by: Scott Rifenbark <scott.m.rifenbark@intel.com>