adt-manual: Updated the build toolchain section with more detail.
Fixes [YOCTO #8368]

Apparently there is confusion on how to build a cross-toolchain
using Eclipse and the Yocto plug-in.  Specifically on setting
the MACHINE variable, the bblayers.conf stuff, and the
preferences in Eclipse to find things.  I added a couple key
things to the section that describes using the OpenEmbedded
build system to build out the toolchain.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>