ref-manual: Updated configuration of auto.conf in closer look
Fixes [YOCTO #8941]

Confusing wording existed surrounding how the auto.conf file
needs to be created.  Early in the section indicates that you
must hand create the file.  Then later, it indicates to not
hand-create it.  In reality, this file is not auto-created by
any script but is really usually created by an autobuilder.
I updated the verbiage to better explain this.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>