sdk-manual: Added note for running remote apps with SSH port forw enabled.
I added the following note in the section that deals with running
Eclipse with a remote application.

Fixes [YOCTO #9276]

Currently, Eclipse does not support SSH port forwarding.
Consequently, if you need to run or debug a remote
application using the host display, you must create a
tunneling connection from outside Eclipse and keep
that connection alive during your work.
For example, in a new terminal, run the following:

     ssh -XY user_name@remote_host_ip

After running the command, add the command to be executed
in Eclipse's run configuration before the application
as follows:

     export DISPLAY=:10.0

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>