sdk-manual: Removed notes about the BitBake Commander
Fixes [YOCTO #10032]

The fix to remove the BitBake Commander from the tools to install
when dealing with the Eclipse Yocto Plug-in has happened.
Consequently, I removed the developer notes in the manual stating
that the Commander was still showing up there.  The manual set
is now clean of this stuff.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>