ref-manual, dev-manual: Add information on devpyshell
Fixes [YOCTO #9166]

In the dev-manual, I added a new section titled "Using a
Development Python Shell."  This section is similar to the
section that talks about using devshell.  The section is
tied to a reference section on the do_devpyshell task.

In the ref-manual, I entered a new task reference for the
do_devpyshell task.  It is brief and references into the
new section in the dev-manual.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>