yocto-project-qs: Altered MinnowBoard MAX example
Fixes [YOCTO #9667]

The example that built the image out for the MinnowBoard MAX was
buiding a core-image-minimal.  This was not ideal.  I have fixed
it so that several types of images are suggested as examples with
a reference to the Images chapter in the ref-manual.  The actual
command now builds out core-image-base.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>