ref-manual: Updated the USERADD_ERROR_DYNAMIC and USERADDEXTENSION variables.
The value for USERADD_ERROR_DYNAMIC changed from "1" to "error"
to activate this variable.  Also, the value "warn" is used
to cause a warning to be issued.  I updated the variable
description to reflect this.

For the USERADDEXTENSION variable, I updated the [doc] string
to use proper quote marks.  I was not nesting these properly.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>