ref-manual: Updated the RDEPENDS variable description.
Fixes [YOCTO #10445]

Pointed out that the EXTENDPKGV variable is hard to locate in the
BB manual.  Firstly, we don't document that variable in the
BB manual.... maybe we should.  For the ref-manual, however, it is
barely referenced in a side example from the dev-manual.  So, I
updated the RDEPENDS variable to have a "Tip" box in the area where
it is shown how to use the package version as part of the RDEPENDs
variable.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>