ref-manual, dev-manual: Added references to SYSROOT_DIRS
Fixes [YOCTO #10692]

I updated the STAGING_DIR variable description to have a reference
to the SYSROOT_DIRS variable as well as the "Sharing Files
Between Recipes" section of the dev-manual.

I also updated the "Sharing Files Between Recipes" section so
that the role of SYSROOT_DIRS is explained and actually used
in an example to add another directory.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>