dev-manual: Added section on bmap-tool to flash images
Fixes [YOCTO #10621]

bmaptool is integrated into the OpenEmbedded build system but is
not documented.  I added a new section describing how to flash an
image to media using the tool.  Also, updated a small section in the
Wic part of the manual that used "dd" to flash an example.  I added
a bmaptool counterpart here and referenced the reader back to the
main new section.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>