ref-manual: Added information about pre-installed dependencies
Fixes [YOCTO #10350]

A situation exists where a user deploys a new application using
devtool target-deploy and the new application has runtime dependencies
that are not already on the target.  When this situation exists, the
application misbehaves if it calls a missing function.  I added the
explanation of this situation to the section that describes the
devtool deploy-target command.  Also, put a pointer in the section
on devtool add for cross-reference purposes.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>