dev-manual, ref-manual: Added support for WKS_FILE_DEPENDS variable.
Fixes [YOCTO #11017]

There is no current way to specify or list plug-in dependencies
when building a Wic image.  The variable WKS_FILE_DEPENDS was introduced
to allow the user to use this in the recipe building the image.
I added a description and example of the variable to the ref-manual
variable glossary.  I also cross-referenced the variable in the
"Plug-ins" section of the dev-manual to point to the variable for
more information.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>